Refactor/cw721 (#862)
