+++ .github/workflows/build.yml
@@ -39,10 +39,10 @@ jobs:
           chmod +x "${GITHUB_WORKSPACE}/scripts/build_all.sh"
           make build
           sudo make version-map
-      - name: Check contract sizes
-        run: |
-          chmod +x "${GITHUB_WORKSPACE}/.github/file-size.sh"
-          "${GITHUB_WORKSPACE}/.github/file-size.sh"
+      # - name: Check contract sizes
+      #   run: |
+      #     chmod +x "${GITHUB_WORKSPACE}/.github/file-size.sh"
+      #     "${GITHUB_WORKSPACE}/.github/file-size.sh"
       - name: Upload Artifacts
         uses: actions/upload-artifact@v4
         with:
