+++ contracts/non-fungible-tokens/andromeda-cw721/Cargo.toml
@@ -26,7 +26,6 @@ testing = ["cw-multi-test", "andromeda-testing"]
 cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw721-base = { workspace = true }
 cw721 = { workspace = true }
