+++ packages/andromeda-non-fungible-tokens/src/cw721.rs
@@ -2,10 +2,15 @@ use andromeda_std::error::ContractError;
 use andromeda_std::{amp::addresses::AndrAddr, andr_exec, andr_instantiate, andr_query};
 use cosmwasm_schema::{cw_serde, QueryResponses};
 
-use cosmwasm_std::{Binary, Coin, CustomMsg, DepsMut, Empty, Env, MessageInfo, Response};
-use cw721::{CollectionExtension, Expiration, RoyaltyInfo};
+use cosmwasm_std::{Binary, Coin, CustomMsg, Deps, DepsMut, Env, MessageInfo, Response};
 
-use cw721_base::{msg::ExecuteMsg as Cw721ExecuteMsg, msg::QueryMsg as Cw721QueryMsg};
+use cw721::msg::{ApprovalsResponse, OperatorsResponse, OwnerOfResponse, TokensResponse};
+use cw721::Expiration;
+
+use cw721_base::{
+    msg::ExecuteMsg as Cw721ExecuteMsg, msg::InstantiateMsg as BaseInstantiateMsg,
+    msg::QueryMsg as Cw721QueryMsg, traits::Cw721Query, Cw721BaseContract,
+};
 
 #[andr_instantiate]
 #[cw_serde]
@@ -20,18 +25,35 @@ pub struct InstantiateMsg {
     pub minter: AndrAddr,
 }
 
-use cw721::traits::{Cw721Execute, Cw721Query};
+use cw721::traits::Cw721Execute;
 
 #[derive(Default)]
-pub struct AndrCW721Contract;
-impl Cw721Execute<Empty, Empty, Empty, Empty, Empty, Empty> for AndrCW721Contract {}
-impl Cw721Query<Option<Empty>, Option<CollectionExtension<RoyaltyInfo>>, Empty>
-    for AndrCW721Contract
-{
+pub struct AndrCW721Contract {
+    standard_implementation: Cw721BaseContract<'static>,
 }
 
 // Add a custom query method to handle the conversion
 impl AndrCW721Contract {
+    pub fn new() -> Self {
+        Self {
+            standard_implementation: Cw721BaseContract::default(),
+        }
+    }
+
+    // Entry points
+    pub fn instantiate(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        msg: InstantiateMsg,
+    ) -> Result<Response, ContractError> {
+        let standard_msg = convert_instantiate_msg(msg);
+        self.standard_implementation
+            .instantiate(deps, env, info, standard_msg)
+            .map_err(|e| e.into())
+    }
+
     pub fn execute(
         &self,
         deps: DepsMut,
@@ -39,54 +61,153 @@ impl AndrCW721Contract {
         info: &MessageInfo,
         msg: Cw721ExecuteMsg,
     ) -> Result<Response, ContractError> {
-        // Convert the message to a compatible type for the trait implementation
-        let cw721_msg = match msg {
-            Cw721ExecuteMsg::TransferNft {
-                recipient,
-                token_id,
-            } => cw721::msg::Cw721ExecuteMsg::TransferNft {
-                recipient,
-                token_id,
-            },
-            Cw721ExecuteMsg::SendNft {
-                contract,
-                token_id,
-                msg,
-            } => cw721::msg::Cw721ExecuteMsg::SendNft {
-                contract,
-                token_id,
-                msg,
-            },
-            Cw721ExecuteMsg::Approve {
-                spender,
-                token_id,
-                expires,
-            } => cw721::msg::Cw721ExecuteMsg::Approve {
-                spender,
-                token_id,
-                expires,
-            },
-            Cw721ExecuteMsg::Revoke { spender, token_id } => {
-                cw721::msg::Cw721ExecuteMsg::Revoke { spender, token_id }
-            }
-            Cw721ExecuteMsg::ApproveAll { operator, expires } => {
-                cw721::msg::Cw721ExecuteMsg::ApproveAll { operator, expires }
-            }
-            Cw721ExecuteMsg::RevokeAll { operator } => {
-                cw721::msg::Cw721ExecuteMsg::RevokeAll { operator }
-            }
-            Cw721ExecuteMsg::Burn { token_id } => cw721::msg::Cw721ExecuteMsg::Burn { token_id },
-            // Add other conversions as needed
-            _ => return Err(ContractError::new("Unsupported execute message")),
-        };
+        self.standard_implementation
+            .execute(deps, env, info, msg)
+            .map_err(|e| e.into())
+    }
 
-        // Call the trait implementation's execute method and convert the error type
-        match <Self as Cw721Execute<Empty, Empty, Empty, Empty, Empty, Empty>>::execute(
-            self, deps, env, info, cw721_msg,
-        ) {
-            Ok(response) => Ok(response),
-            Err(_) => Err(ContractError::new("Error executing CW721 command")),
-        }
+    pub fn transfer_nft(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        recipient: String,
+        token_id: String,
+    ) -> Result<Response, ContractError> {
+        self.standard_implementation
+            .transfer_nft(deps, env, info, recipient, token_id)
+            .map_err(|e| e.into())
+    }
+
+    pub fn revoke_all(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        token_id: String,
+    ) -> Result<Response, ContractError> {
+        self.standard_implementation
+            .revoke_all(deps, env, info, token_id)
+            .map_err(|e| e.into())
+    }
+
+    pub fn send_nft(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        contract_addr: String,
+        token_id: String,
+        msg: Binary,
+    ) -> Result<Response, ContractError> {
+        self.standard_implementation
+            .send_nft(deps, env, info, contract_addr, token_id, msg)
+            .map_err(|e| e.into())
+    }
+
+    pub fn burn_nft(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        token_id: String,
+    ) -> Result<Response, ContractError> {
+        self.standard_implementation
+            .burn_nft(deps, env, info, token_id)
+            .map_err(|e| ContractError::new(&e.to_string()))
+    }
+
+    pub fn approve(
+        &self,
+        deps: DepsMut,
+        env: &Env,
+        info: &MessageInfo,
+        spender: String,
+        token_id: String,
+        expires: Option<Expiration>,
+    ) -> Result<Response, ContractError> {
+        self.standard_implementation
+            .approve(deps, env, info, spender, token_id, expires)
+            .map_err(|e| ContractError::new(&e.to_string()))
+    }
+
+    pub fn query(
+        &self,
+        deps: Deps,
+        env: &Env,
+        msg: Cw721QueryMsg,
+    ) -> Result<Binary, ContractError> {
+        self.standard_implementation
+            .query(deps, env, msg)
+            .map_err(|e| e.into())
+    }
+
+    pub fn query_owner_of(
+        &self,
+        deps: Deps,
+        env: &Env,
+        token_id: String,
+        include_expired_approval: bool,
+    ) -> Result<OwnerOfResponse, ContractError> {
+        self.standard_implementation
+            .query_owner_of(deps, env, token_id, include_expired_approval)
+            .map_err(|e| e.into())
+    }
+
+    pub fn query_approvals(
+        &self,
+        deps: Deps,
+        env: &Env,
+        token_id: String,
+        include_expired_approval: bool,
+    ) -> Result<ApprovalsResponse, ContractError> {
+        self.standard_implementation
+            .query_approvals(deps, env, token_id, include_expired_approval)
+            .map_err(|e| e.into())
+    }
+
+    pub fn query_operators(
+        &self,
+        deps: Deps,
+        env: &Env,
+        owner: String,
+        include_expired_approval: bool,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    ) -> Result<OperatorsResponse, ContractError> {
+        self.standard_implementation
+            .query_operators(
+                deps,
+                env,
+                owner,
+                include_expired_approval,
+                start_after,
+                limit,
+            )
+            .map_err(|e| e.into())
+    }
+
+    pub fn query_all_tokens(
+        &self,
+        deps: Deps,
+        env: &Env,
+        start_after: Option<String>,
+        limit: Option<u32>,
+    ) -> Result<TokensResponse, ContractError> {
+        self.standard_implementation
+            .query_all_tokens(deps, env, start_after, limit)
+            .map_err(|e| e.into())
+    }
+}
+
+fn convert_instantiate_msg(msg: InstantiateMsg) -> BaseInstantiateMsg {
+    BaseInstantiateMsg {
+        name: msg.name,
+        symbol: msg.symbol,
+        minter: Some(msg.minter.into_string()),
+        collection_info_extension: None,
+        creator: None,
+        withdraw_address: None,
     }
 }
